Project code/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
streamlit>=1.28.0
plotly>=5.17.0
Pillow>=10.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0

Flask>=2.2.0
